Make bounded lattice construction implicit
Summary:
Security lattice used to explicitly add bottom and top policies to the input lattice, but this is not very useful for checking legal flows from bottom to a purpose (or from a purpose to top) created on-the-fly (as opposed to explicitly mentioned in the lattice input).

So we delay the boundedness to check time to see if the flow being validated contains top or bottom.

This fixes the superfluous reporting of PUBLIC to PURPOSE privacy errors.

Reviewed By: zilberstein

Differential Revision: D22791752

fbshipit-source-id: 9743cceffba7970266938c7f28672e8600731956