Avoid rebuilding exn tree
Summary:
I'd like to modify HHBBC's analysis phase to operated on a const Func*. There's one mutating operation that we do here, but I don't think we need to do it.

If we really wanted to keep this optimization around, we could update do_optimize to call it instead - but I just don't think it's needed. It's just a block-level DCE on exception blocks.

Reviewed By: ricklavoie

Differential Revision: D22809979

fbshipit-source-id: 66a51abc45144b7c6db2f695d571b5eb1e085ee0