Ban inst_meth equivalent syntax
Summary:
After discussion, we have decided to ban `inst_meth` equivalent syntax. That is: we will not be allowing

`$foo->bar<...>` as syntax that is able to get a function pointer. Users can achieve similar functionality to this via writing their own closure.

Part of the reasoning for not allowing this vs toplevel functions and class methods is because toplevel functions and class methods both do not need to capture objects and therefore in HHVM can be better optimized. inst_meth will always capture an object.

Reviewed By: Wilfred

Differential Revision: D21942142

fbshipit-source-id: cff340adc0b3a15c7d9283dd008cf425d1283808