Typecheck reified function pointers
Summary:
We want to ensure that when a function pointer is created to a function with reified generics, that we correctly capture those types when passing around the function pointer.

`reified_check` is modified. Normally it expects that any call of a function with reified generics has the generics bound at the call site. Since function pointers break this assumption, we want to ensure that we check the reified generics for function pointers but do not need to check them when calling it. So added a bit in the flag map to mark function pointers.

Reviewed By: Wilfred

Differential Revision: D22080890

fbshipit-source-id: 14ede11594f6b23afee93cb30a76bb137c1ae234