propagate exception constraints across functions
Summary:
When you call a function that may throw an exception, it can leak information about that function's program counter. We want to take this into account in a similar way to how we take PCs into account when calling other functions.

To address this, I've added another policy to the function prototype. We add dependencies from the exceptions that are thrown to this policy and when a function gets called, its exception policy gets added to the current PC for Next and Catch.

Reviewed By: madgen

Differential Revision: D22597536

fbshipit-source-id: ba66eb0b9e1dff28f8f9934d67573dcea56bd2cb