Don't crash when someone tries to new an [Async]Generator.
Summary:
Both Generator and AsyncGenerator are among the classes in systemlib that try to prevent themselves from being constructed by user code by having a private constructor. They also both have native data, and unlike other examples that have both (ext_hsl_io, ext_icu_* and ext_async_mysql) there is no function in the same class that news them (which allows calling the private constructor) and there is no legit use case for making mocks of them that ignore the normal constructor (as is the case with ext_hsl_io).

The default constructor for both of these classes produces instances that can't safely be used or even destroyed, so currently if you try to new one of them: In normal code (not the debugger) we allocate the instance (in an invalid state) then try to call the constructor and throw due to it being private, then segfault while trying to destroy the instance on the stack while we propagate the exception. In the debugger the privacy check is skipped, so we end up with an instance that you can try to call methods on (which will also segfault).

Fix this in a way that's similar to what we do for Awaitables: make the class' instanceCtor throw, which prevents the instance from being allocated in the first place, so the only way to create instances is in native functions that take a different code path to allocate them.

I intially wanted to do this by giving them a null instanceInit in their NativeDataInfo and `= delete` ing their default constructors, but unfortunately this doesn't currently work because when we clone an object with natove data we make a copy of the native data by first default constructing an instance at the destination and then copying over it with operator=. We can do that as a follow-up if we change the mechanism for cloning native data to use a copy constructor rather than default construction followed by operator= (this requires refactoring all classes with native data so I'm not doing it on this bugfix diff).

Instead, for now I add a separate bool flag to NativeDataInfo to indicate that rather than installing Native::nativeDataInstanceCtor as the instanceCtor in the class we should install a throwingInstanceCtor (similar to asioInstanceCtor) that immediately throws an exception when called.

Add unit tests to make sure we don't break this again.

Reviewed By: ricklavoie

Differential Revision: D22833753

fbshipit-source-id: 65474ad9f72b94d8d1e683ceed0361be37c03bbb