Use lowptr for FCall context
Summary:
I'm pretty sure this string is a low string in lowptr builds, just like the SA immediate type. It's decoded from the unit the same way.

While we're at it, there's a small inaccuracy in the FCallArgs move constructor: the uniqueness criteria only applies to the heap-allocated FCallArgsLong, not to FCallArgsShort.

Finally, we can also delete some more unused macros here.

Reviewed By: jano

Differential Revision: D22876483

fbshipit-source-id: 970861e363487020dcb0a45c5c348da97788a3ee