solved some TODOs about Tgeneric type arguments (1)
Summary: When adding type arguments to Typing_defs_core.Tgeneric to implement higher-kinded types, I left a lot of TODOs in places where we needed to handle the newly added type arguments. This diff just removes some of those TODOs in places where I've convinced myself that the type arguments can safely be ignored. This is the first of two similar diffs.

Reviewed By: vsiles

Differential Revision: D22665446

fbshipit-source-id: cae442242aba483aa45a63565a8b91a8735a5eb6