Unstable Feature Flagging
Summary:
Currently today the only way to try out experimental features in Hack is to change the .hh_config file. This requires hh_server to restart, which can be painful for large enough code bases. In some cases these features also requires changing the hhvm config as well. This can be a tedious process if all a developer wants to do is test out an experimental feature for a bit.

This diff adds the capability of enabling experimental features within a single source file by using the special attribute `__EnableUnstableFeatures`. By passing in the name of the feature as a string literal, it will enable the feature within the desired scope. For example:

```
<<file:__EnableUnstableFeatures('foo')>>
...
```
Would enable the feature `foo` within the file. This allows us to only retype check the given file instead of having to do a full restart of hh_server. The other advantage of it being in the file is HackC has access to this information so extra gating does not need to be done via hhvm configs.

Right now this is best used for features that introduce new syntactic constructs. We can then looks for these constructs in `rust_parser_errors.rs` and report an error if an unstable feature is utilized.

In order to run unstable features the hhvm config Hack.Lang.AllowUnstableFeatures must be enabled. Otherwise presence of the special attribute will be reported as an error. There is an analogous setting in .hhconfig for hh_server

Note: this is my first Rust diff so open to feedback

Reviewed By: losvald

Differential Revision: D22838676

fbshipit-source-id: a4f03b6c784b61129654115fa253646debf3d7c3