Use binary distribution of OPAM
We originally didn't do this due to the concerns similar to those raised
in the classic 'Reflections on Trusting Trust' paper.

As Rust and Cargo do not support a similar process, we've effectively
lost that battle in the typechecker now - so, let's take the same
approach for the ocaml stuff.

Previously, we would:
1. Download the ocaml source
2. Build ocaml
3. Download the opam source
4. Use our new shiny ocaml to build opam
5. Use opam to build the ocaml we actually want for Hack, and all the
dependencies

Now we have:
1. Fetch the precompiled binary distribution of opam
2. Use opam to build the ocaml we actually want for Hack, and all the
dependencies

This is mostly to speed up build times, especially on diff-time CI
within Facebook.