Suggest cleaning `hh_fanout` state on state-load failure
Summary:
During development, loading the state from disk sometimes fails, because we unmarshal an OCaml blob that was intended for a previous version of the executable. Add a `clean` command and a hint to use it when this scenario occurs.

So far, I've only ever seen this with the "development" version of `hh_fanout`, not for any deployed version. The state path is keyed by the version of `hh_fanout` (and the development version is keyed by the string "development" instead of a version hash), so we don't have to worry about this scenario in production. This is just for future maintainers, and for when I don't want to manually figure out what path I need to `rm -rf` to get things working again.

Reviewed By: ljw1004

Differential Revision: D22678095

fbshipit-source-id: 65f3fcd7022ef17a5caea7a37f733de15f9a0d12