Move check for override attribute into typing
Summary:
Currently, if a method has an `<<__Override>>` attribute but doesn't override anything, we produce an error at decl time (if the method is in a class) or typing time (if the method is from a trait).

Let's move both to typing - we want to eliminate error production from decl.

Reviewed By: vsiles

Differential Revision: D22900703

fbshipit-source-id: 7548cedff1eef2bd018d9a2cadc1da1abdd3cb61