inline methods in serverMain
Summary:
See tip of stack D22757723

This is a trivial refactoring diff. I'm inlining the method `update_recheck_values`. That's because I'm aiming for a more holistic treatment of logging, and abstracting the logs behind a function makes it hard to follow.

I also renamed `recheck_loop` into `recheck_until_no_changes_left`. That's to avoid confusion with the other main server loop.

Reviewed By: arxanas

Differential Revision: D22689361

fbshipit-source-id: b6a2db99c2c875e32ccafdb282540eec8cadcb22