log init_telemetry
Summary:
See D22757733. It explains the motive for "init_telemetry".

This diff completes that vision. It makes it so the SERVER_IS_READY event, and the corresponding server log entry, both have full insight into (1) the saved-state load telemetry about dirty files, (2) the typecheck telemetry about how long it took to do the many phases of typechecking

I also reworded Hh_logger.log so it's clear which HackEventLogger event it's referring to.

Reviewed By: arxanas

Differential Revision: D22757723

fbshipit-source-id: 9e2e6f9b80fbcf005a080e6239f2fe80f492a409