pass start_time to type_check_core
Summary:
When we print "Total: 2.3 seconds", what does that total time represent? Is it the time since we started recheck_loop? Or the time after we'd finished some preliminaries and started typechecking? or the time we started type_check_core?

I decided that the Telemetry.t object should uniformally log all times as offsets from the start of recheck_loop.

This diff changes type_check_core to do that. (It used to be measuring time since the start of type_check_core.) I figure that uniformity is the most important thing.

Reviewed By: arxanas

Differential Revision: D22699860

fbshipit-source-id: 434c505b462191fe37122993ee0147a6a72e093c