refactor ServerUtils.log_hash_stats
Summary:
For all typechecker telemetry, I adopted a compositional pattern where each function produces its own telemetry object. That makes for telemetry that's a bit harder to read, but I think it's the best way to deal with telemetry that comes from a huge tree of nested functions.

The `ServerUtils.log_hash_stats` function didn't follow that pattern. It took a telemetry object and added fields to it.

I'm switching to the other pattern, for uniformity.

Reviewed By: 2BitSalute

Differential Revision: D22757731

fbshipit-source-id: 07e900ac379c9edcfb83497b1128e8cf55ef0ff7