cleanup of diag_sub logging
Summary:
This is a trivial refactoring diff. I kept looking at this body of code in serverMain, which I'd written last year to track down a bug, and thinking that surely it must be possible to make it look nicer.

I also changed it so the server log prints "RECHECK_END" instead of "Recheck id:". My idea is that this way it's easier for us to understand that the event we see in the server log is the same as the event we see in HackEventLogger telemetry.

Reviewed By: arxanas

Differential Revision: D22689371

fbshipit-source-id: d9577478047c97769e8df2a94ff2d8661a8e4924