"hh --json" receives Telemetry.t
Summary:
When you do `hh --json` then hh_client receives telemetry about the typecheck, so it can emit this telemetry as part of the json. Prior to this diff, it conveys the information in `ServerCommandTypes.Recheck_stats.t`

My project is for the server to gather much richer telemetry than it currently does, and to gather it into a Telemetry.t object. This object is a write-only abstraction, in the sense that only producers of telemetry know the structure of what they're going to produce, and clients merely dump out the entire opaque telemetry blob.

In this diff I change the server so it no longer sends `Recheck_stats.t`. Instead it sends a `Telemetry.t` object, and hh_client has no knowledge of the inner structure of this, and it merely prints whatever telemetry the server has told it.

There's really no need for the client to have knowledge about the telemetry structures that the server uses.

Reviewed By: arxanas

Differential Revision: D22689360

fbshipit-source-id: 558063c12ee977bf840a92bf58b7d0ce137045d2