rename to batch_telemetry[]
Summary:
Typechecking has a loop called "recheck_until_no_changes_left". It's job is to look for file-changes, typecheck, look for more file-changes, typecheck again, and keep going until there are no more file-changes. Each iteration is called a "batch".

We used to store telemetry in the form `{batch_1:{...}, batch_2:{...}, ...}`

That feels wrong. This diff changes it to `[{...}, {...}]`

Reviewed By: arxanas

Differential Revision: D22689370

fbshipit-source-id: d22be108164017b4efe2b5a68e2ada0f1ea9420b