architecture of init_telemetry
Summary:
I want to be able to relate what we find upon load-saved-state (number of dirty master files, dirty local files) to typechecking time.

This is how it works:
1. We only know about dirty files in the `HackEventLogger.type_check_dirty` event. This merely stores the list of files-that-need-typechecking inside `env`, so that a subsequent iteration of the main server loop can pick up those files and typecheck based on them.
2. The main server loop will run a "recheck loop" which looks for file-changes, typechecks everything that must be checked (including the full init prechecked work and the fanout), looks for more file-changes, processes these, and so on until there are no more file changes. This logs in `HackEventLogger.recheck_end`. This is the only place that knows about the totality of all batches.
3. Within each iteration (also called "batch") it calls into `HackEventLogger.type_check_end`. This is the only place that knows whether it was interrupted by file-changes.
4. The first iteration to complete the saved-state-init-fanout calls `HackEventLogger.server_is_ready`. This is the only place that knows when the init is done.

I want to correlate the information, so for instance I can relate the number of dirty files to the time it takes to typecheck their fanout. I'd like to be able to read this directly from our telemetry, rather than having to sql-join.

This diff introduces an architecture for it. It makes it so the save-state-loader stores telemetry about itself (including dirty files). This telemetry is kept around until the typechecker comes to process the dirty files, and then we can print out full contextual information about both the origin and the performance of the typecheck.

The actual implementation is pretty trivial, in serverEnv.ml:
```
BEFORE:
(* Whether a full check was ever completed since init. *)
needs_full_init: bool;

AFTER:
(* Whether a full check was ever completed since init, and why it was needed *)
why_needs_full_init: Telemetry.t option;
```
We already threaded through "needs_full_init" from exactly the right place in serverLazyInit, through to the place that performs the typecheck. This makes it a low-risk way to achieve my goals.

Reviewed By: arxanas

Differential Revision: D22757733

fbshipit-source-id: 8c1add7a43a6d6b397a0de5a47a9969d2917dcaa