Stop using submodules for fb-mysql and zstd
This should speed up builds for pretty much everyone: the tarball is
< 150mb, compared to > 550mb for the submodule.

Also, it's only downloaded if it's needed - for example, it's not needed
if we're only building Hack.

If you're reading this commit message because you want to update it,
reach out to the FB MySQL team to push an updated tag if needed.