Fix segmentation fault when using invalid Thrift spec for the second time
Summary:
I noticed that the one of the Thrift tests consistently segfaulted when run more than once. This turned out to be due to the Thrift spec cache, which is used to avoid compiling specs more than once. In the case that a static spec was compiled unsuccessfully (and threw an exception), an empty entry would be left in the cache map, leading to a segfault when it was used a second time.

To reproduce: `hphp/test/run --retranslate-all 2 hphp/test/slow/thrift/bad_spec.php`

I have updated the test to validate this behavior when run normally, and fixed the cache to remove the new entry if compilation throws.

If you would prefer me to refactor `compileSpec` to no longer throw exceptions (and pass this responsibility up to its caller), let me know and I can do that instead.

Reviewed By: mofarrell

Differential Revision: D22922851

fbshipit-source-id: f9475f8e859499064f5da249bef9910d6eabfb00