Add SlabAllocator
Summary:
The values we store in cachelib are ocamlrep Slabs--contiguous, movable byte strings each containing a tree of OCaml values. We currently write a decl to cachelib by allocating an OCaml value in an ocamlrep::Arena via ToOcamlRep, then copying that value into a Slab byte string by traversing the value tree (in `ocamlrep::slab::OwnedSlab::from_value`), then memcpying that byte string into cachelib-managed memory. The tree traversal we use to construct the slab is expensive--I've measured it taking as long as the ToOcamlRep conversion took in the first place (both conversion-to-OCaml and the slab-building traversal take roughly 25% of the time it takes to parse the decl). It would be much cheaper if the Arena built a slab in the first place--then we could skip the traversal and simply memcpy the slab into cachelib.

Now that ToOcamlRep::to_ocamlrep returns an OpaqueValue, this is possible. We can write a slab-oriented arena which uses a Vec as its backing memory, and whose allocation functions return offsets into that vec rather than absolute memory addresses. This requires a few changes to BlockBuilder and the Allocator trait--it's no longer possible for consumers to assume that a BlockBuilder represents an absolute memory address, and helpers like bytes_to_ocamlrep, which would like the absolute memory address of a block, must ask the Allocator to convert a BlockBuilder to a pointer (just in case the BlockBuilder contains an offset rather than a pointer).

Reviewed By: arxanas

Differential Revision: D22361505

fbshipit-source-id: 07eb40871abb157e774d5640b38ff6658aa5c363