Don't show dependent path in `hh_fanout debug`
Summary:
The `dependent_path` field here is misleading. It sounds like it's the file that the dependent symbol resides in.

It's not. It's actually the file that we happened to be typechecking when we discovered the dependency in question. This is oftentimes the file that the dependent symbol resides in, but not always. For example, in a method call chain, we may end up adding edges from a retrieved symbol (for an expression of inferred type) to the symbols it depend on.

See the next diff for an example of a test case where we add dependencies for symbols in files other than the one we're typechecking.

Differential Revision: D22831636

fbshipit-source-id: 3164d0139ed92156ca35b67fb62fd2765c6aa4cb