Adds a server config flag for using the highlighted error formatter by default
Summary:
This diff adds a server config flag for using the highlighted error formatter by default, which will automatically be set based on the `cpe_hh_highlighted_error_formatter` gatekeeper (found [here](cpe_hh_highlighted_error_formatter)).

An overarching goal of this change was to not need to check anything on the client (including gatekeeper, as it takes around 1 second to do so). Instead, the server checks the configs in `.hhconfig`/`hh.conf`/gatekeeper and sends a flag over RPC to the client indicating whether the highlighted error formatter should be used.

This change is meant to be temporary while we roll-out the new formatter and get feedback. I imagine in a future diff we'll roll-back all these changes and just set the error_format default to be Highlighted in `hh_client`.

I've left `hh_lite` the same; we can update that in a later diff.

Reviewed By: Wilfred

Differential Revision: D22871835

fbshipit-source-id: ded4a97e685b1128f76cba5a23dfd4e2757643a1