Add default toDynamic and toString for IsTypeStructProfile and SwitchProfile
Summary:
Noticed that not all of the target profiles implement these overloads. Since where we want to trace doesn't always have the size information that these two profiles expect, then it's not possible to even build.

In this case, let's just return empty strings. We're not planning on understanding these specific profiles yet so this is safe to do.

Differential Revision: D22834668

fbshipit-source-id: f936f398b40cf2cf8c3290292ddd6bfce6438d23