Use hashtbl for deduping in linearization
Summary:
Our current means of avoiding emitting duplicate elements in a linearization is needlessly expensive. We "normalize" the element we wish to emit by reallocating it with positions replaced by Pos.none, then iterate over the list of already-emitted elements, normalizing each and comparing it with the normalized candidate element.

Instead of iterating over the entire list of emitted elements (which makes linearization take quadratic time), we could iterate over only the emitted elements with the same class name (which is likely to be a very small list) if we mantain a hash table of emitted elements keyed by class name. We can avoid paying for normalization by manually writing a comparison routine that ignores positions.

Reviewed By: arxanas

Differential Revision: D22929292

fbshipit-source-id: 1f1b0fddc2f8cc2624c61bda5d64acfecd9bf157