Don't poll for progress if --from is supplied
Summary:
Reading through rage logs, what strikes me is (1) there are a huge number of `--from hackast` invocations, (2) the monitor is continually bombarded with loads of progress requests, indeed so much so that the logs are hard to read and it's hard to see what is the cause of "monitor busy".

I've also heard anecdotally that people invoking hackast a lot find more "monitor busy" than usual.

This diff adds a new boolean argument `hh_client check --show-progress <true|false>`. If it's true, then clientCheck will send progress requests to the monitor once a second. If it's false, then clientCheck won't. If it's absent, then it defaults to true normally, but it defaults to false if you provided a `--from` argument.

My rationale is that the folks who do a client check with `--from` are almost all scripts, and they never have any need for progress polling.

My goal with this diff is to have easier rage logs to investigate. And, fingers crossed, maybe this will do something to reduce the likelihood of monitor busy.

Differential Revision: D22938356

fbshipit-source-id: 35676fa301f34ec7ba9c0ee7b8d6ee1c3e3065b5