Implement basic bytecode compression
Summary:
In this diff, we introduce a new set of compression helpers that can flatten an HHBBC BytecodeVec into a flat array of bytes (suitable, potentially, for further compression). Flattening bytecode is not trivial. Some bytecode immediates have references to other heap structures, and some of these references are even owned (i.e. when we delete the old bytecode, the target would be destructed).

In this diff, we solve this problem by handling each op in a way that's aware of the types of its inputs. We'd like to take advantage of existing meta-programming helpers and to do so in a safe and clear way. We end up implementing templated encode and decode helpers that switch on their type parameter. The correspondence between the two is clear so it should be easy to keep them in sync; in addition, they treat broad swathes of types homogeneously (e.g. all trivially destructible types are copied as bytes), so we shouldn't need to updated them too often.

It took me some time to ensure correctness for these helpers, so I ended up writing a bunch of tools to help with that. First off, the tracing for these helpers is quite good: at trace level 4, it clearly displays each bytecode and each immediate type being decoded or encoded, plus offsets, in an output format that will match exactly between the two ops. That made it easy to track down all my initial discrepancies. I also installed a new "--test-compression" HHBBC flag, so that we can do a large-scale full-repo test; that caught a few more bugs specific to FCallArgs.

The new tracing also displays the overall compression ratio. Currently, flattening bytecode is already reduces that memory by 2.75x. This ratio is an understatement, as we don't count any auxiliary heap allocations as part of the original bytecode's memory. There's another easy win to be had here before we use real compression algorithms: int immediates are highly compressible (and indeed, that's why we shift things like BlockIds by NoBlockId so that they're all >= 0).

Reviewed By: mofarrell

Differential Revision: D22861369

fbshipit-source-id: 21150020fe240d3dac748cc0d463d69fe5a360dc