One neat trick can improve compression by 1.5x!
Summary:
In the previous diff, we were careful to encode things like "valid BlockId || NoBlockId" by shifting the value so that the invalid ID was 0. In this diff, we take advantage of this fact by encoding uint32_t in 1 byte (if it's < 254), in 3 bytes (if it's < (1 << 16)), or 5 bytes otherwise.

On systemlib, the new encoding increases our compression ratio from 2.85x to 4.37x. For comparison, if we leave out the 16-bit case, we only improve the ratio to 3.78x.

Reviewed By: mofarrell

Differential Revision: D22970434

fbshipit-source-id: c4016834ff7b0f883c59668b709e49c0ed815aa3