Fix ordering of error reasons when their respective positions aren't in the same file.
Summary:
The order of reasons presented (e.g. "Expected foo", "Got bar") was sometimes getting mixed up depending on whether they were adjacent and occurred in the same file (i.e. adjacent reasons in the error list that weren't in the same file might not be presented in the correct order).

This diff fixes it by remembering the original index of each marked message, as well as by turning the marked message tuples into real records (with the happy consequence of cleaning up the code a little). I think adding the original index into the record, though only used in a handful of places, may be useful in the future for other things.

This fix also illuminated another test file that had been reporting the reasons in the wrong order (`array_map_explicitly_typed_error.php`), so this is also fixed.

Reviewed By: Wilfred

Differential Revision: D22861091

fbshipit-source-id: 7d7bc2f6edb2528e8ea81328256fed7c89274060