ServerCommandTypes.finale_data -> Exit.finale_data
Summary:
This diff removes the old ServerCommandTypes.finale_data and replaces it with the new Exit.finale_data.

The two are largely similar. The only difference is that Exit.finale_data has an optional string user-facing msg field (since we produce this finale_data in all cases), whereas it used to be mandatory.

I've updated the user-facing msg displays in clientLsp and clientConnect accordingly, to deal with None msg.

Note: changing it from mandatory to optional will change the binary file format of the finale file. This is okay, because the finale file is only written by the server and read by the client after they've already agreed their version handshake.

I had to remove "Exit" module from Lsp. That's because "open Lsp" caused module "Exit" to be shadowed by "Lsp.Exit".

Reviewed By: arxanas

Differential Revision: D22935085

fbshipit-source-id: 632dafadc38b4e03cbed7b55248bc6cce5e3420c