Move GlobalConfig.default_sharedmem_config into SharedMem.default_config
Summary:
This diff, though trivial, is a bit of a slog. I wrote it to break up the build dependency
* GlobalConfig.tmp_file -> SharedMem -> EventLogger

It just didn't seem right that in order to get hold of the temp file, you needed to drag in EventLogger.

It also meant that EventLogger was unable to know what the tmp directory is, which is a shame. (and something I needed for the next diffs up the stack)

Reviewed By: arxanas

Differential Revision: D22931195

fbshipit-source-id: 59df8367d75bdce15a234133c8403454b764767a