Exit.exit
Summary:
This diff creates a new `Exit.exit` method which will (next diff) replace `Exit_status.exit`

That's because I want to put more machinery in there. I want it so that every time the server does a deliberate exit, then it writes a "finale" file, so that the client can know why it exited.

That would be too much machinery to go into `Exit_status.exit`, because Exit_status us part of utils/core:core, which is the very tiniest most minimal set of files.

Reviewed By: 2BitSalute

Differential Revision: D22931446

fbshipit-source-id: 2e17a56b339f1f136b70dd14f52eded1eddfd64d