Server exit always writes finale file
Summary:
With this diff, if the server was launched in non-check mode, and if it exits via an Exit.exit statement, then it will always write a finale file.

It used to be that only one single place wrote a finale file (in ServerInit, when it couldn't load a saved-state). I've changed this codesite to use the now standard way of writing the finale file.

Reviewed By: arxanas

Differential Revision: D22935141

fbshipit-source-id: 1ee9c63b9b315ac016fa63a020fd7330e1a96abd