support implicit policies via entailment
Summary:
In order to support functions that are governed by implicit policies, we need to use entailment. More precisely, if we inferred some set of constraints `C` by analyzing a function with prototype `fp`, then it must be the case that:

```
fp <: cipp_proto |= C
```

In this case, cipp_proto is a prototype where the implicit policy flows into all input and the pc, and exceptions and the return also flow into cipp. This essentially means that all the inputs, output, and pc are governed by the same policy.

Reviewed By: madgen

Differential Revision: D22818981

fbshipit-source-id: 28e970d77d0e6d31927d491df42a961a56e8fe59