defer entailment check to after solving
Summary: The goal of this diff is to enable cipp function to be able to call other functions. Currently this is not possible because the entailment check fails when we encounter holes (which are inserted for function calls). To fix the issue, I defer the entailment check until the holes are closed. This allows us to call other cipp functions as well as InferFlows functions (as long as they do not violate the cipp invariants).

Reviewed By: madgen

Differential Revision: D22926556

fbshipit-source-id: ea9b37280db6cefe2a89e9dcea4abe2de1d2e46a