Const-correctness for Func pointers
Summary:
This diff is pure code-motion. We clean up most places where we violate const-correctness for Func pointers.
- Context can just include a const Func* now
- Emit now takes a const Func*; to renumber locals, we create a new Locals array
- We no longer delete bytecode in emit; we already do it in whole-program
- We store a const Func* for a class's methods in the index rather than a Func*

The big remaining whole is that AnalysisContext still includes a MutFunc. I think I'll solve that by moving the context out of FuncAnalysisResult, so that the caller has to supply it along with the result, and for the optimize pass, we'll have a context with a MutFunc instead. It will be a bit involved, so I'll keep it separate.

Reviewed By: arnabde03

Differential Revision: D22818523

fbshipit-source-id: 1fbce437b9488e204cee906d61db96e4f6d9d854