Align heap slots on word instead of cacheline boundaries
Summary:
Currently, we align heap slots in shared memory on cacheline boundaries i.e. multiples of 64 bytes. This results in quite a bit of wasted space to pad out entries (around 4% of the total).

Aligning on word boundaries (8 bytes) recovers most of this space, with no apparent effect on performance.

Differential Revision: D23011387

fbshipit-source-id: cbc49df05d6636976b890b5f82acbc2e82df8ce9