Modify subtyping for shapes in case of undefined fields
Summary:
The previous subtyping logic would look up shape fields and give the type `nothing` for an undefined shape field. My earlier draft of this commit kept the rest of the logic, and simply changed the undefined lookup to give a non-value using an option type, with extra validity checks.

Now I have refactored the subtyping logic to be more principled instead, following a draft document on intersection types.

Reviewed By: CatherineGasnier

Differential Revision: D22868098

fbshipit-source-id: 8f1fb90d846d7afa9f691cf62ff36705abb19e36